requirements_cc.txt
absl-py==0.11.0
alembic==1.9.4+computecanada
appdirs==1.4.4+computecanada
astroid==2.14.2
autopep8==1.5.5
black==23.1.0
bottle==0.12.25
cachetools==5.3.0+computecanada
certifi==2021.5.30+computecanada
charset_normalizer==2.0.6+computecanada
click==8.0.1+computecanada
cmaes==0.9.1+computecanada
colorlog==6.7.0+computecanada
configparser==5.0.2+computecanada
cycler==0.10.0+computecanada
decorator==4.4.2+computecanada
dill==0.3.6+computecanada
docker-pycreds==0.4.0+computecanada
easyocr==1.2.1
flake8==5.0.4
gitdb==4.0.7+computecanada
GitPython==3.1.24+computecanada
google-api-core==2.11.0
google-auth==2.16.2
google-cloud-vision==3.4.0
googleapis-common-protos==1.58.0
greenlet==1.1.0+computecanada
grpcio==1.38.1+computecanada
grpcio-status==1.33.2
idna==3.2+computecanada
imageio==2.9.0
importlib-metadata==3.4.0
importlib_resources==5.12.0+computecanada
isort==5.11.5
joblib==1.2.0+computecanada
kiwisolver==1.3.1+computecanada
lazy-object-proxy==1.6.0+computecanada
Mako==1.2.4+computecanada
Markdown==3.3.3
MarkupSafe==2.0.1+computecanada
matplotlib==3.1.1+computecanada
mccabe==0.7.0
mypy-extensions==1.0.0
networkx==2.5+computecanada
numpy==1.21.0+computecanada
olefile==0.46
opencv-python==4.5.1.48+computecanada
optuna==3.1.0+computecanada
optuna-dashboard==0.9.0
packaging==23.0+computecanada
pandas==1.3.0+computecanada
pathspec==0.11.0+computecanada
pathtools==0.1.2+computecanada
Pillow==7.0.0+computecanada
Pillow-SIMD==7.0.0.post3+computecanada
piqa==1.2.2
platformdirs==3.0.0+computecanada
promise==2.3+computecanada
proto_plus==1.22.2+computecanada
protobuf==3.20.3
psutil==5.8.0+computecanada
pyasn1==0.4.8+computecanada
pyasn1-modules==0.2.8+computecanada
pycodestyle==2.9.1
pyflakes==2.5.0
pylint==2.16.2
pyparsing==2.4.7+computecanada
python-bidi==0.4.2
python-dateutil==2.8.1+computecanada
python-Levenshtein==0.12.0+computecanada
pytz==2021.3+computecanada
PyWavelets==1.1.1+computecanada
PyYAML==5.3.1+computecanada
requests==2.26.0+computecanada
rsa==4.9+computecanada
scikit-image==0.18.1+computecanada
scikit_learn==1.0.1+computecanada
scipy==1.7.0+computecanada
sentry-sdk==1.3.1+computecanada
setproctitle==1.3.2
shortuuid==1.0.1+computecanada
six==1.15.0+computecanada
smmap==4.0.0+computecanada
SQLAlchemy==1.4.21+computecanada
subprocess32==3.5.4+computecanada
tensorboard==1.15.0
termcolor==1.1.0+computecanada
tesserocr==2.5.1
threadpoolctl==3.1.0+computecanada
tifffile==2021.2.1
toml==0.10.1
tomli==2.0.1+computecanada
tomlkit==0.11.6
torch==1.8.0+computecanada
torchvision==0.11.1+computecanada
tornado==6.1+computecanada
tqdm==4.61.2+computecanada
typed-ast==1.4.3+computecanada
typing_extensions==4.5.0+computecanada
Unidecode==1.1.1
urllib3==1.26.7+computecanada
wandb==0.13.9+computecanada
Werkzeug==1.0.1
wrapt==1.15.0
yaspin==2.1.0+computecanada
zipp==3.4.0